refactored, configured to build with firebase, added bloc to pubspec and main.dart
