+++ thoughts/ios/Runner/GoogleService-Info.plist
@@ -19,15 +19,15 @@
 	<key>STORAGE_BUCKET</key>
 	<string>thoughts-53463.appspot.com</string>
 	<key>IS_ADS_ENABLED</key>
-	<false></false>
+	<false/>
 	<key>IS_ANALYTICS_ENABLED</key>
-	<false></false>
+	<false/>
 	<key>IS_APPINVITE_ENABLED</key>
-	<true></true>
+	<true/>
 	<key>IS_GCM_ENABLED</key>
-	<true></true>
+	<true/>
 	<key>IS_SIGNIN_ENABLED</key>
-	<true></true>
+	<true/>
 	<key>GOOGLE_APP_ID</key>
 	<string>1:367708514211:ios:5628d7334bf5720b3a5c16</string>
 </dict>
