+++ thoughts/lib/constants/colors.dart
