+++ thoughts/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:thoughts/colors.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:thoughts/constants/colors.dart';
 import 'package:thoughts/screens/homeScreen.dart';
 
 void main() {
@@ -9,16 +10,19 @@ void main() {
 class ThoughtsApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primaryColor: themeColor,
-        splashColor: Colors.transparent, // reduces splashes from the app
-        highlightColor: Colors.transparent, // reduces highlights from the app
-        visualDensity: VisualDensity.adaptivePlatformDensity,
+    return MultiBlocProvider(
+      providers: [],
+      child: MaterialApp(
+        title: 'Flutter Demo',
+        theme: ThemeData(
+          primaryColor: themeColor,
+          splashColor: Colors.transparent, // reduces splashes from the app
+          highlightColor: Colors.transparent, // reduces highlights from the app
+          visualDensity: VisualDensity.adaptivePlatformDensity,
+        ),
+        home: HomeScreen(),
+        debugShowCheckedModeBanner: false,
       ),
-      home: HomeScreen(),
-      debugShowCheckedModeBanner: false,
     );
   }
 }
